==> 01. Data - preparing.ipynb <==
# Figuring out what data we want.
The *turtle* parent category has WnID *n01662784*, see http://www.image-net.org/synset?wnid=n01662784

import pandas as pd

WnID_parent_child = pd.read_csv('./data/WnID_parent_child_relations.txt', header=None, sep=' ').rename(columns={0: 'parent', 1: 'child'})

turtle_children = WnID_parent_child.loc[WnID_parent_child['parent'] == 'n01662784']
turtle_grandchildren = WnID_parent_child.loc[WnID_parent_child['parent'].isin(
    turtle_children['child'])].rename(columns={'parent': 'child', 'child': 'grandchild'})
turtle_grandgrandchildren = WnID_parent_child.loc[WnID_parent_child['parent'].isin(
    turtle_grandchildren['grandchild'])].rename(columns={'parent': 'grandchild', 'child': 'grandgrandchild'})

print('There are no grand-grand-grand-children:')
turtle_grandgrandgrandchildren = WnID_parent_child.loc[WnID_parent_child['parent'].isin(
    turtle_grandgrandchildren['grandgrandchild'])].rename(columns={'parent': 'grandgrandchild', 'child': 'grandgrandgrandchild'})
display(turtle_grandgrandgrandchildren)

turtle_lineage = pd.merge(pd.merge(turtle_children, turtle_grandchildren, on='child', how='left'), 
                          turtle_grandgrandchildren, on='grandchild', how='left')
display(turtle_lineage.set_index(['parent', 'child', 'grandchild', 'grandgrandchild']))

For parents with a single child, we decided to go with the parent class. An example for this is the WnID *n01667778*: this refers to *Terrapin* (http://www.image-net.org/synset?wnid=n01667778) which has only the *Diamondback Terrapin* as a child. This will also be useful for the bounding box data since if boxes are available they are available for the parent in such cases, but not the child.

kept_WnIDs = turtle_lineage['grandgrandchild']
kept_WnIDs.loc[~turtle_lineage['grandchild'].duplicated(keep=False)] = turtle_lineage['grandchild'][~turtle_lineage['grandchild'].duplicated(keep=False)]
kept_WnIDs.loc[~turtle_lineage['child'].duplicated(keep=False)] = turtle_lineage['child'][~turtle_lineage['child'].duplicated(keep=False)]

Now we can take a look at the associated species.

WnID_imagenet_class_relation = pd.read_csv(
    './data/WnID_to_classname.txt', header=None, sep='	').rename(columns={0: 'WnID', 1: 'Label'}).set_index('WnID')
WnID_turtle_species_relation = WnID_imagenet_class_relation.loc[kept_WnIDs.values]
WnID_turtle_species_relation = WnID_turtle_species_relation['Label'].str.split(',', expand=True)[0].str.title()
turtle_information = WnID_turtle_species_relation.rename('Species').reset_index().set_index('Species')

print('We have data for {} species.'.format(len(turtle_information)))
display(turtle_information)

We downloaded the tar files for these WnIDs. The files are provided in ``./data/archives/images``.

## Now we look at bounding boxes

with open('./data/WnID_bounding_boxes.txt', 'r') as f:
    bounding_box_WnIDs = [line.strip() for line in f.readlines() if line.strip() in kept_WnIDs.values]

turtle_information['bboxes'] = False
turtle_information.loc[turtle_information['WnID'].isin(bounding_box_WnIDs), 'bboxes'] = True

print('We have bounding boxes for {} species'.format(len(bounding_box_WnIDs)))
print('Specifically, the following species.')
display(turtle_information.loc[turtle_information['bboxes']])

We downloaded the tar.gz files for these bounding boxes. The files are provided in ``./data/archives/bounding_boxes``.

# Train-test-validation split and file extraction

import tarfile
import os

Before extracting the files, we put together a tabel of what we got.

image_level_information = list()
for img_archive in os.listdir('./data/archives/images'):
    with tarfile.open('./data/archives/images/' + img_archive) as tar:
        for member in tar:
            image_level_information.append(member.name)

bbox_information = list()
for img_archive in os.listdir('./data/archives/bounding_boxes'):
    with tarfile.open('./data/archives/bounding_boxes/' + img_archive) as tar:
        for member in tar:
            if member.name.endswith('.xml'):
                bbox_information.append(member.name)
            else:
                continue

print('We have a total of {} images and bounding box data for {} images.'.format(
    len(image_level_information), len(bbox_information)))

image_level_statistics = pd.DataFrame(image_level_information, columns=['img_name'])
image_level_statistics = pd.DataFrame(image_level_statistics['img_name'].str.rstrip('.JPEG'))
image_level_statistics['WnID'] = image_level_statistics['img_name'].str.split('_', expand=True)[0]
image_level_statistics.set_index('img_name', inplace=True)
img_with_bbox = pd.Series(bbox_information).str.split('/', expand=True)[2].str.rstrip('.xml').str.strip().values
image_level_statistics['bboxes'] = False
image_level_statistics.loc[image_level_statistics.index.isin(img_with_bbox), 'bboxes'] = True

print('We do not have an image for every bounding box. We could make the connection for {} images.'.format(
    image_level_statistics['bboxes'].sum()))

# merge the image count per species
turtle_information = pd.merge(turtle_information.reset_index(), 
                              image_level_statistics.groupby('WnID').count().rename(columns={'bboxes': 'n_images'}).reset_index(),
                              on='WnID')
# mergeturtle_information the count of images with bboxes per species
turtle_information = pd.merge(turtle_information, 
                             image_level_statistics.groupby('WnID').sum().rename(columns={'bboxes': 'n_bbox_imgs'}).reset_index(),
                             on='WnID')

turtle_information.set_index('Species', inplace=True)

print('Number of images per species')
display(turtle_information['n_images'])
print('------\n')
print('Number of images with bounding boxes')
display(turtle_information.loc[turtle_information['bboxes'],'n_bbox_imgs'].astype(int))

## Now we decide for each image whether it is supposed to go into train, validation, or test set.
Since the final evaluation metric is top-1 and top-3 error, we do not need bounding boxes in the test set. Since bounding boxes are not only not for all species, but also not for all images within a species available, this is helpful. 

We are aiming want a stratified 60-20-20 split across all images. Meanwhile, the bounding box images should have a 80-20-0 split.

image_level_statistics['sample'] = 'train'

First we do the split stratified by bounding boxes. Only have to set 20% to validation here within species.

for wnid in bounding_box_WnIDs:
    seed = 91420171 # date and first seed.
    filt = image_level_statistics['bboxes'] & (image_level_statistics['WnID'] == wnid)
    draw = image_level_statistics.loc[filt].sample(frac=0.2, random_state=seed).index
    image_level_statistics.loc[draw, 'sample']= 'val'

Next, we have to take 20% of the images per species and put them in the test sample, making sure that we only draw from those that are not annotated

for wnid in image_level_statistics['WnID'].unique():
    # Now we take the train data for which we don't have boxes and split it to get an overall 60% train 
    # and 20% test set. Thus, the test set is 25% of the train set(0.25*0.8 = 0.2) corresponding to
    # 0.25 * n_total_train_images. We take test images only from the not annotated images, so 
    # the proportion from these is 0.25 * n_total_train_images / n_not_annotated_train_images
    seed = 91420172
    filt_non_bb = ~image_level_statistics['bboxes'] 
    filt_wnid = (image_level_statistics['WnID'] == wnid)
    frac = 0.2 * (filt_wnid.sum()/ (filt_wnid & filt_non_bb).sum())
    draw = image_level_statistics.loc[filt_non_bb & filt_wnid].sample(frac=frac, random_state=seed).index
    image_level_statistics.loc[draw, 'sample']= 'test'
    
    # Now we have to balance the validation set so that we have a 60-20-20 split while maintaining the 
    # 80-20-0 split for the bounding box annotated images. We first count the number of images already
    # in the validation set and then adjust.
    seed = 91420173
    filt_training = (image_level_statistics['sample'] == 'train')
    filt_validation = (image_level_statistics['sample'] == 'val')
    n_add_to_validation = int(0.2 * filt_wnid.sum() - (filt_wnid & filt_validation).sum())
    draw = image_level_statistics.loc[filt_non_bb & filt_wnid & filt_training].sample(
        n=n_add_to_validation, random_state=seed).index
    image_level_statistics.loc[draw, 'sample']= 'val'

Now we have to check that this all worked.

print('The 80-20 training validation split for the bounding box data worked well')
pd.concat([
    image_level_statistics.loc[
        image_level_statistics['bboxes'] & (image_level_statistics['sample'] == 'train')
    ].groupby('WnID')['sample'].count().rename('training'),
    image_level_statistics.loc[
        image_level_statistics['bboxes'] & (image_level_statistics['sample'] == 'val')
    ].groupby('WnID')['sample'].count().rename('validation')
], axis=1
).divide(
    image_level_statistics.loc[
        image_level_statistics['bboxes']].groupby('WnID')['sample'].count(), axis='index'
).rename(
    columns={})

print('...and so did the overall 60-20-20 split')
pd.concat([
    image_level_statistics.loc[
        (image_level_statistics['sample'] == 'train')
    ].groupby('WnID')['sample'].count().rename('training'),
    image_level_statistics.loc[
        (image_level_statistics['sample'] == 'val')
    ].groupby('WnID')['sample'].count().rename('validation'),
    image_level_statistics.loc[
        (image_level_statistics['sample'] == 'test')
    ].groupby('WnID')['sample'].count().rename('test'),
], axis=1
).divide(
    image_level_statistics.groupby('WnID')['sample'].count(), axis='index'
).rename(
    columns={})

With this, we can move on to extract the files into their appropriate destinations.

os.mkdir('./data/images/train')
os.mkdir('./data/images/val')
os.mkdir('./data/images/test')

def get_sample(member):
    return image_level_statistics.loc[member.name.rstrip('.JPEG'), 'sample']

for tar_file in os.listdir('./data/archives/images'):
    wnid = tar_file.rstrip('.tar')
    os.mkdir(os.path.join('./data/images/train', wnid))
    os.mkdir(os.path.join('./data/images/val', wnid))
    os.mkdir(os.path.join('./data/images/test', wnid))
    with tarfile.open(os.path.join('./data/archives/images',tar_file)) as tar:
        for member in tar:
            tar.extract(member, 
                        os.path.join('./data/images', get_sample(member), wnid))

Quick check that this worked

import numpy as np

for wnid in image_level_statistics['WnID'].unique():
    folder_size = np.array(
        [len(os.listdir('./data/images/train/' + wnid)),
         len(os.listdir('./data/images/val/' + wnid)),
         len(os.listdir('./data/images/test/' + wnid))])
    print(wnid)
    print(folder_size/np.sum(folder_size))

Looks good!

Finally, extract the bounding box data

os.mkdir('./data/bounding_boxes/train')
os.mkdir('./data/bounding_boxes/val')

def get_sample(img_name):
    return image_level_statistics.loc[img_name, 'sample']

for tar_file in os.listdir('./data/archives/bounding_boxes'):
    wnid = tar_file.rstrip('.tar.gz')
    os.mkdir(os.path.join('./data/bounding_boxes/train', wnid))
    os.mkdir(os.path.join('./data/bounding_boxes/val', wnid))
    with tarfile.open(os.path.join('./data/archives/bounding_boxes',tar_file)) as tar:
        for member in tar:
            img_name = member.name.split('/')[-1].rstrip('.xml')
            member.name = img_name + '.xml'
            if img_name in image_level_statistics.index:
                tar.extract(member, 
                            os.path.join('./data/bounding_boxes', get_sample(img_name), wnid))
            else:
                continue

Check whether this worked as well

for wnid in image_level_statistics.loc[image_level_statistics['bboxes'],'WnID'].unique():
    folder_size = np.array(
        [len(os.listdir('./data/bounding_boxes/train/' + wnid)),
         len(os.listdir('./data/bounding_boxes/val/' + wnid))])
    print(wnid)
    print(folder_size/np.sum(folder_size))

Nice :)

# Get some more statistic for the images.

### Check for broken data

from PIL import Image

import warnings
broken_data = dict()

for sample in ('train', 'val', 'test'):
    for wnid in image_level_statistics['WnID'].unique():
        path = os.path.join('data', 'images', sample, wnid)
        imgs = os.listdir(path)
        for img in imgs:
            with warnings.catch_warnings():
                warnings.filterwarnings('error')
                try:
                    im = Image.open(os.path.join(path,img))
                except Warning as e:
                    broken_data[os.path.join(path,img)] = e

print('We have {} broken images.'.format(len(broken_data)))
print(broken_data)

Since we have so few broken images, we are just going to delete them.

for path in broken_data.keys():
    img_name = path.split('/')[-1].rstrip('.JPEG')
    image_level_statistics.drop(img_name, inplace=True)
    os.remove(path)
    print('Dropped and deleted {}'.format(img_name))

### Look at image resolutions

image_level_statistics['pixel_w'] = None
image_level_statistics['pixel_h'] = None

def get_resolution(sample, wnid, img_name):
    path = os.path.join('./data/images/', sample, wnid, img_name + '.JPEG')
    im = Image.open(path)
    return(im.size)

for row in image_level_statistics.reset_index().iterrows():
    row = row[1] #dont want the index
    image_level_statistics.loc[
        row['img_name'],['pixel_w', 'pixel_h']
    ] = get_resolution(row['sample'], row['WnID'], row['img_name'])

image_level_statistics[['pixel_w', 'pixel_h']] = image_level_statistics[
    ['pixel_w', 'pixel_h']].astype(int)

print('Descriptives width and height')
image_level_statistics[['pixel_w', 'pixel_h']].describe().astype(int)

Do the same by species.

# append species
image_level_statistics = pd.merge(
    image_level_statistics.reset_index(), 
    turtle_information.reset_index()[['WnID', 'Species']],
    how='left', on='WnID')

image_level_statistics.groupby('Species')[
    ['pixel_w', 'pixel_h']].describe().astype(int)

# Read the bounding box information from the xml files

import xml.etree.ElementTree as ET

def get_box_coordinates_from_xml(wnid, sample, img_name):
    xml_path = os.path.join('./data/bounding_boxes/', sample, wnid,
                            img_name + '.xml') 
    tree = ET.parse(xml_path)
    objects = tree.findall('object')
    boxes = []
    for object in objects:
        bbox = object.find('bndbox')
        box_coords = [str(child.text) for child in bbox]
        boxes.append(box_coords)
    return boxes

bounding_boxes_by_img = image_level_statistics.loc[
    image_level_statistics['bboxes']].apply(
    lambda row: get_box_coordinates_from_xml(row['WnID'], 
                                             row['sample'],
                                             row['img_name']), 
    axis=1).rename('box_coords')

print('We encounter the following number of boxes per image:')
print(bounding_boxes_by_img.apply(len).unique())

image_level_statistics = pd.concat(
    [image_level_statistics, bounding_boxes_by_img], axis=1)

image_level_statistics = pd.concat(
    [image_level_statistics,
     image_level_statistics.loc[image_level_statistics['bboxes'],  
                                'box_coords'].apply(len).rename('n_boxes')], 
    axis=1)

print('Let\'s look at the number of boxes by species')
image_level_statistics[image_level_statistics['bboxes']].groupby('Species')['n_boxes'].describe()

There are few images with more than one box: the 75% for each species is still a single box!

print('Looking at this in more detail, we have the following number of images with more than one box.')
image_level_statistics[image_level_statistics['n_boxes'] > 1].groupby('Species')['n_boxes'].count()

print('The total number of boxes available is {}'.format(
    int(image_level_statistics['n_boxes'].sum())))

image_level_statistics.to_csv(
    './results/data_analysis/image_level_statistics.csv')

# Finally, we prepare a txt file that we use for training the F-RCNN
This matches the structure for the *simple-parser*: `filepath,x1,y1,x2,y2,class_name,imageset`.
The first stage trains merely for the distinction of turtletoise vs background. Thus, we only have a single class: *turtletoise*!

simple_parse = []
for row in image_level_statistics[
    image_level_statistics['bboxes']].iterrows():
    sample = row[1]['sample']
    wnid = row[1]['WnID']
    img_file = row[1]['img_name'] + '.JPEG'
    boxes = row[1]['box_coords']
    filepath = os.path.join('./data/images/', sample, wnid,
                           img_file)
    for box in boxes:
        simple_parse.append(','.join([filepath, ','.join(box),
                                       'turtletoise', sample]) + '\n')

with open('./f_rcnn_simple_parser_first_stage.txt', 'w') as f:
    f.writelines(simple_parse)

# Looking at a specific example for the train-validation-test split

import pandas as pd
image_level_statistics = pd.read_csv('./results/data_analysis/image_level_statistics.csv')

leatherback_n_imgs = image_level_statistics.loc[image_level_statistics.Species == 'Leatherback Turtle', 'bboxes'].count()
leatherback_n_bbox_imgs = image_level_statistics.loc[(image_level_statistics.Species == 'Leatherback Turtle') 
                                                 & image_level_statistics.bboxes, 'bboxes'].count()

print('For Leatherbacks, we have {} images total, {} of which also have bounding boxes'.format(leatherback_n_imgs, leatherback_n_bbox_imgs))

leatherback_train_test_val = image_level_statistics[
    image_level_statistics.Species == 'Leatherback Turtle'].groupby('sample')['bboxes'].agg(['count', 'sum'])

leatherback_train_test_val

print('In percent')
pd.concat([leatherback_train_test_val['count']/leatherback_n_imgs,
           leatherback_train_test_val['sum']/leatherback_n_bbox_imgs], axis=1) * 100   

leatherback_n_imgs - leatherback_n_bbox_imgs - 194

582-332
