==> 03. Benchmark - setup and fit.ipynb <==
from keras.applications.vgg16 import VGG16, preprocess_input
from keras.preprocessing.image import ImageDataGenerator
import numpy as np

base_vgg16 = VGG16(weights='imagenet')

# Benchmark 1: VGG16 with replaced last layer

from keras.models import Model
from keras.layers import Dense

Replace the layer

final_layer = Dense(22, activation='softmax')(base_vgg16.get_layer('fc2').output)
benchmark_last_layer = Model(base_vgg16.input, final_layer)

benchmark_last_layer.summary()

Make all other layers not trainable

for layer in benchmark_last_layer.layers[:-1]:
    layer.trainable = False

from keras.metrics import top_k_categorical_accuracy
def top_1_error(y_true, y_pred):
    return 1-top_k_categorical_accuracy(y_true, y_pred, k=1)

def top_3_error(y_true, y_pred):
    return 1-top_k_categorical_accuracy(y_true, y_pred, k=3)

benchmark_last_layer.compile(loss='categorical_crossentropy', optimizer='adam',
                            metrics=[top_1_error, top_3_error]) 

Data generators for preprocessed data

def preprocess_img(img):
    img = np.expand_dims(img, axis=0)
    return preprocess_input(img)

datagen_train = ImageDataGenerator(preprocessing_function=preprocess_img,
                                   horizontal_flip=True)

def generator_train(batch_size):
    return datagen_train.flow_from_directory(
        './data/images/train', target_size=(224, 224),
        batch_size=batch_size, class_mode='categorical')

datagen_val = ImageDataGenerator(preprocessing_function=preprocess_img)

def generator_val(batch_size, shuffle=True):
    return datagen_val.flow_from_directory(
        './data/images/val', target_size=(224, 224),
        batch_size=batch_size, class_mode='categorical', 
        shuffle=shuffle)

import pandas as pd
image_level_statistics = pd.read_csv('./results/data_analysis/image_level_statistics.csv')
n_train_images = (image_level_statistics['sample'] == 'train').sum()
n_val_images = (image_level_statistics['sample'] == 'val').sum()

Some callbacks. We save the model with the lowest validation loss, stop if there was no decrease in 20 epochs, and track the history of losses and errors.

from keras.callbacks import ModelCheckpoint, EarlyStopping, LambdaCallback
checkpointer = ModelCheckpoint(filepath='./saved_models/benchmark/weights_vgg16_top_only.hdf5', 
                               verbose=1, save_best_only=True)
stopper = EarlyStopping(monitor='val_loss', min_delta=0, patience=20, verbose=1, mode='auto')

history_last_layer = list()
def get_history(history_dict):
    historian = LambdaCallback(
        on_epoch_end=lambda epoch, logs:  history_dict.append(
            {'loss': logs['loss'], 
             'top_1_error': logs['top_1_error'],
             'top_3_error': logs['top_3_error'], 
             'val_loss': logs['val_loss'],  
             'val_top_1_error': logs['val_top_1_error'], 
             'val_top_3_error': logs['val_top_3_error']}
        )
    )
    return historian

epochs = 100

batch_size = 32

benchmark_last_layer.fit_generator(generator_train(batch_size), 
                                   steps_per_epoch=n_train_images/(5*batch_size),
                                   epochs=epochs, 
                                   callbacks = [checkpointer, stopper, 
                                                get_history(history_last_layer)], 
                                   validation_data=generator_val(batch_size), 
                                   validation_steps=n_val_images/(5*batch_size), 
                                   verbose=2)

benchmark_last_layer.load_weights('./saved_models/benchmark/weights_vgg16_top_only.hdf5')
performance_benchmark_last_layer = benchmark_last_layer.evaluate_generator(
    generator_val(batch_size, shuffle=False), steps = n_val_images/batch_size)
print('Validation loss: {:2f}'.format(performance_benchmark_last_layer[0]))
print('Validation top-1-error rate: {:.2f}%'.format(100*performance_benchmark_last_layer[1]))
print('Validation top-3-error rate: {:.2f}%'.format(100*performance_benchmark_last_layer[2]))

Save the results

pd.DataFrame(history_last_layer).to_csv('./results/benchmark/history_last_layer.csv')

# Benchmark 2: VGG16 with replaced dense layers
Looking at the losses from Benchmark 1, the model is probably not flexible enough but still quickly over-fitting. Lets try to change things up. We exchange the flatten layer for a global average pooling layer and add our own dense layers after that. (We cannot directly re-use the initial layers since the first Dense layer is just too big for our compouter.) We add dropout. The new dense layers are initiated with random weights. 

from keras.layers import Dropout, GlobalAveragePooling2D

dense = GlobalAveragePooling2D()(base_vgg16.get_layer('block5_pool').output)
dense = Dense(2048, activation='relu', name='fc1')(dense)
dense = Dropout(0.3)(dense)
dense = Dense(2048, activation='relu', name='fc2')(dense)
dense = Dropout(0.3)(dense)
dense = Dense(22, activation='softmax', name='prediction')(dense)

benchmark_all_dense_layers = Model(base_vgg16.input, dense)

benchmark_all_dense_layers.summary()

print('Which layers are trainable?')
for layer in benchmark_all_dense_layers.layers:
    print('{}: {}'.format(layer.name, layer.trainable))

benchmark_all_dense_layers.compile(loss='categorical_crossentropy', optimizer='adam',
                                   metrics=[top_1_error, top_3_error])

checkpointer = ModelCheckpoint(filepath='./saved_models/benchmark/weights_vgg16_dense.hdf5', 
                               verbose=1, save_best_only=True)

history_all_dense_layers = list()

batch_size=25

benchmark_all_dense_layers.fit_generator(generator_train(batch_size), 
                                         steps_per_epoch=n_train_images/(5*batch_size),
                                         epochs=epochs, 
                                         callbacks = [checkpointer, stopper, 
                                                      get_history(history_all_dense_layers)], 
                                         validation_data=generator_val(batch_size), 
                                         validation_steps=n_val_images/(5*batch_size), 
                                         verbose=2)

benchmark_all_dense_layers.load_weights('./saved_models/benchmark/weights_vgg16_dense.hdf5')
performance_benchmark_all_dense_layers = benchmark_all_dense_layers.evaluate_generator(
    generator_val(batch_size, shuffle=False), steps = n_val_images/batch_size)
print('Validation loss: {:2f}'.format(performance_benchmark_all_dense_layers[0]))
print('Validation top-1-error rate: {:.2f}%'.format(100*performance_benchmark_all_dense_layers[1]))
print('Validation top-3-error rate: {:.2f}%'.format(100*performance_benchmark_all_dense_layers[2]))

Save results

pd.DataFrame(history_all_dense_layers).to_csv('./results/benchmark/history_all_dense_layers.csv')

# Benchmark 3: VGG16 with replaced dense layers and trained convolutional block
Benchmark 2 is doing better on the validation data. Let's try to be even more flexible by allowing the last convolutional block to train as well. Here we are going to use the weights from Benchmark 2 to initialize. We just use the same model and make some more layers trainable. 

benchmark_dense_and_conv = benchmark_all_dense_layers

for layer in benchmark_dense_and_conv.layers[15:]:
    layer.trainable = True

print('Which layers are trainable?')
for layer in benchmark_dense_and_conv.layers:
    print('{}: {}'.format(layer.name, layer.trainable))

We use a slower learning rate here for fine-tuning

from keras.optimizers import SGD
sgd = SGD(lr=10e-4, momentum=0.9, nesterov=True)

benchmark_dense_and_conv.compile(loss='categorical_crossentropy', optimizer=sgd,
                                 metrics=[top_1_error, top_3_error])

checkpointer = ModelCheckpoint(filepath='./saved_models/benchmark/weights_vgg16_conv_and_dense.hdf5', 
                               verbose=1, save_best_only=True)

history_dense_and_conv = list()

batch_size=20

benchmark_dense_and_conv.fit_generator(generator_train(batch_size), 
                                       steps_per_epoch=n_train_images/(5*batch_size),
                                       epochs=epochs, 
                                       callbacks = [checkpointer, stopper, 
                                                    get_history(history_dense_and_conv)], 
                                       validation_data=generator_val(batch_size), 
                                       validation_steps=n_val_images/(5*batch_size), 
                                       verbose=2)

benchmark_dense_and_conv.load_weights('./saved_models/benchmark/weights_vgg16_conv_and_dense.hdf5')
performance_benchmark_conv_and_dense_layers = benchmark_dense_and_conv.evaluate_generator(
    generator_val(batch_size, shuffle=False), steps = n_val_images/batch_size)
print('Validation loss: {:2f}'.format(performance_benchmark_conv_and_dense_layers[0]))
print('Validation top-1-error rate: {:.2f}%'.format(100*performance_benchmark_conv_and_dense_layers[1]))
print('Validation top-3-error rate: {:.2f}%'.format(100*performance_benchmark_conv_and_dense_layers[2]))

Doing better than the previous models :).

Save results

pd.DataFrame(history_dense_and_conv).to_csv('./results/benchmark/history_dense_and_conv.csv')

from keras.models import load_model
load_model('./saved_models/benchmark/weights_vgg16_conv_and_dense.hdf5', compile=False).summary()
