==> 05. First stage - setup and fit.ipynb <==
# First stage R-CNN
Fit the F-RCNN model to those images for which we have bounding boxes. We do this with binary labels: either turtletoise or not. In doing this, we hope to improve the ability of the model to recognize turtletoises more generally which will be needed for the second step.

## Fit the model

!python ./keras-frcnn/train_frcnn.py --path ./f_rcnn_simple_parser_first_stage.txt --parser simple --network vgg_lite --hf --use_validation True --num_epochs 50 --epoch_length 500 --verbose 2 --input_weight_path ./saved_models/benchmark/weights_vgg16_conv_and_dense.hdf5   --logs_path ./results/f_rcnn/first_stage/first_stage_training.csv --output_weight_path ./saved_models/f_rcnn/first_stage/model_frcnn.hdf5  --logs_path ./results/f_rcnn/first_stage/first_stage_training.csv

## Crop images for second stage

### Train

import os
try:
    os.mkdir('./data/f_rcnn_second_stage/train')
except:
    pass

!python ./keras-frcnn/first_stage_output_frcnn.py -p ./data/images/train --output_path ./data/f_rcnn_second_stage/train --logs_path ./results/f_rcnn/second_stage/crop_imgs_logs_train.csv --num_rois 100 --config_filename ./saved_models/f_rcnn/first_stage/config.pickle --input_weight_path ./saved_models/f_rcnn/first_stage/model_frcnn.hdf5

### Validation

import os
try:
    os.mkdir('./data/f_rcnn_second_stage/val')
except:
    pass

!python ./keras-frcnn/first_stage_output_frcnn.py -p ./data/images/val --output_path ./data/f_rcnn_second_stage/val --logs_path ./results/f_rcnn/second_stage/crop_imgs_logs_val.csv --num_rois 100 --config_filename ./saved_models/f_rcnn/first_stage/config.pickle --input_weight_path ./saved_models/f_rcnn/first_stage/model_frcnn.hdf5

### Test
(we crop now for use later)

import os
try:
    os.mkdir('./data/f_rcnn_second_stage/test')
except:
    pass

!python ./keras-frcnn/first_stage_output_frcnn.py -p ./data/images/test --output_path ./data/f_rcnn_second_stage/test --logs_path ./results/f_rcnn/second_stage/crop_imgs_logs_test.csv --num_rois 100 --config_filename ./saved_models/f_rcnn/first_stage/config.pickle --input_weight_path ./saved_models/f_rcnn/first_stage/model_frcnn.hdf5
